+++ open_mastr/utils/orm.py
@@ -612,8 +612,8 @@ class ElectricityConsumer(ParentAllTables, Base):
     GeplantesInbetriebnahmedatum = Column(Date)
 
 
-class MarketRoles(ParentAllTables, Base):
-    __tablename__ = "market_roles"
+class MarketActorsAndRoles(ParentAllTables, Base):
+    __tablename__ = "market_actors_and_roles"
 
     MastrNummer = Column(String, primary_key=True)
     MarktakteurMastrNummer = Column(String)
@@ -977,14 +977,14 @@ tablename_mapping = {
         "__class__": DeletedMarketActors,
         "replace_column_names": None,
     },
-    "marktrollen": {
-        "__name__": MarketRoles.__tablename__,
-        "__class__": MarketRoles,
-        "replace_column_names": None,
-    },
     "marktakteure": {
         "__name__": MarketActors.__tablename__,
         "__class__": MarketActors,
+        "replace_column_names": None,
+    },
+    "marktakteureundrollen": {
+        "__name__": MarketActorsAndRoles.__tablename__,
+        "__class__": MarketActorsAndRoles,
         "replace_column_names": None,
     },
     "netze": {